% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafilter.R
\name{datafilter}
\alias{datafilter}
\title{Make a data filter}
\description{
Make a data filter

Make a data filter
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{datafilter$new()}}
\item \href{#method-NA-ui}{\code{datafilter$ui()}}
\item \href{#method-NA-make_choices}{\code{datafilter$make_choices()}}
\item \href{#method-NA-apply_filter}{\code{datafilter$apply_filter()}}
\item \href{#method-unknown-clone}{\code{datafilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Make a new filter object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{datafilter$new(
  data,
  column_name,
  label,
  filter_ui = c("picker", "pickersearch", "virtualsearch", "select", "checkboxes",
    "slider", "numeric_range", "numeric_min", "numeric_max", "date_range", "switch"),
  tooltip = "",
  array_field = FALSE,
  array_separator = ";",
  array_comparison = "any",
  search_method = c("equal", "regex"),
  round_digits = 1,
  choices = NULL,
  options = list(),
  ...,
  static = NULL,
  section = NULL,
  updates = NULL,
  server = NULL,
  select_choices = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}

\item{\code{column_name}}{Name column in the dataframe}

\item{\code{label}}{Label (without tooltip!)}

\item{\code{filter_ui}}{Choice of filter UI}

\item{\code{tooltip}}{Tooltip text (optional)}

\item{\code{array_field}}{Logical. If TRUE, the column is coded as an array}

\item{\code{array_separator}}{If array_field, the separator (";" or "json")}

\item{\code{array_comparison}}{Name of function to do array comparison ("any", "all")}

\item{\code{search_method}}{For text-like columns, how to filter ('equal' or 'regex' for pattern search)}

\item{\code{round_digits}}{For numeric-like columns, number of digits to round the filter value to (for numeric range)}

\item{\code{choices}}{Optional, (named) vector of choices for select-like filters}

\item{\code{options}}{Options to the filter, these are settings sent to the actual filter input (e.g. 'width' to selectInput)}

\item{\code{...}}{Further arguments ignored}

\item{\code{static}}{Deprecated}

\item{\code{section}}{Deprecated}

\item{\code{updates}}{Deprecated}

\item{\code{server}}{Deprecated}

\item{\code{select_choices}}{Deprecated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-ui"></a>}}
\if{latex}{\out{\hypertarget{method-NA-ui}{}}}
\subsection{Method \code{ui()}}{
Make the UI for the filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{datafilter$ui(ns = NS(NULL), label = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{Shiny namespace}

\item{\code{label}}{Logical, if FALSE, does not render a label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
For select-input type filters, make the choices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{datafilter$make_choices(x, selected = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of data}

\item{\code{selected}}{Optional vector of pre-selected choices}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-apply_filter"></a>}}
\if{latex}{\out{\hypertarget{method-NA-apply_filter}{}}}
\subsection{Method \code{apply_filter()}}{
Apply the filter to data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{datafilter$apply_filter(data, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}

\item{\code{input}}{Shiny input object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{datafilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
